requirements.txt
google-genai
streamlit
pandas
gspread